(NOT TESTED) Implemented fault recovery of RM and completed promoting mechanisim in case of GS failure. TODO: GS and RM selection.

+++ src/main/java/santiagoAndFerdy/vgs/messages/IDGen.java
@@ -1,9 +1,9 @@
 package santiagoAndFerdy.vgs.messages;
 
 public class IDGen {
-    private static long ID = 0;
+    private static int ID = 0;
 
-    public static long getNewId(){
+    public static int getNewId(){
         return ID++;
     }
 }
